+++ Source/EventFlow.SQLite/Extensions/EventFlowOptionsExtensions.cs
@@ -20,7 +20,6 @@
 // COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-// 
 
 using EventFlow.Configuration;
 using EventFlow.Extensions;
@@ -42,6 +41,7 @@ namespace EventFlow.SQLite.Extensions
             return eventFlowOptions
                 .RegisterServices(f =>
                 {
+                    f.Register<ISQLiteDatabaseMigrator, SQLiteDatabaseMigrator>();
                     f.Register<ISQLiteConnection, SQLiteConnection>();
                     f.Register<ISQLiteConnectionFactory, SQLiteConnectionFactory>();
                     f.Register<ISQLiteErrorRetryStrategy, SQLiteErrorRetryStrategy>();